fix bug and update hra_example
